requirements.txt
amqp==5.1.1
anyio==3.7.0
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asttokens==2.2.1
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
black==23.3.0
bleach==6.0.0
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
channels==4.0.0
charset-normalizer==3.2.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
comm==0.1.3
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.3
cycler==0.11.0
daphne==4.0.0
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dj-database-url==2.0.0
Django==4.2.2
django-bootstrap-v5==1.0.11
django-channels==0.7.0
django-extensions==3.2.3
executing==1.2.0
fastjsonschema==2.17.1
Fiona==1.9.4.post1
flower==2.0.0
fonttools==4.40.0
fqdn==1.5.1
geopandas==0.13.2
humanize==4.7.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
ipykernel==6.24.0
ipython==8.14.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.4
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyter_server==2.7.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
kombu==5.3.1
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==3.0.1
mypy-extensions==1.0.0
nbclassic==1.0.0
nbclient==0.8.0
nbconvert==7.6.0
nbformat==5.9.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.25.0
oauthlib==3.2.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.0.0
platformdirs==3.8.0
prometheus-client==0.17.0
prompt-toolkit==3.0.38
psutil==5.9.5
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
Pygments==2.15.1
pyOpenSSL==23.2.0
pyparsing==3.1.0
pyproj==3.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.8
python-json-logger==2.0.7
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.0
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.2
service-identity==23.1.0
shapely==2.0.1
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.3.2
traitlets==5.9.0
Twisted==22.10.0
txaio==23.1.1
typing_extensions==4.7.0
tzdata==2023.3
uri-template==1.3.0
urllib3==2.0.4
vine==5.0.0
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.1
widgetsnbextension==4.0.7
zope.interface==6.0
